requirements.txt
Pillow==7.1.2
python-telegram-bot==12.8
